FIX(donations): donation management correctly assigns currency_id
The problem with posting donation items originated from donation
management not correctly assigning currency ids.  This was allowed to
occur because the database table does not enforce the currency_id
relation with donation management.

In fact, in the test database, many of the donations have currency_id
set to 0, an invalid currency.  These will need to updated for the
system to work correctly.