FIX(donations): donation management correctly assigns currency_id
The problem with posting donation items originated from donation
management not correctly assigning currency ids.  This was allowed to
occur because the database table does not enforce the currency_id
relation with donation management.

In fact, in the test database, many of the donations have currency_id
set to 0, an invalid currency.  These will need to updated for the
system to work correctly.


Former-commit-id: 6ca84a8c798f20a980937c6b958717980fdda850 [formerly 44cf629ac1a160585fe650d42b3738392209a11c]
Former-commit-id: 28b0b23a0680827487dc86b81e2b63d2941d865b