FIX(find-patient): remove SOUNDEX algorithm; fixes #708
I've updated the patient find directive to use strict matching on
first_name, last_name, middle_name.  We do all possible combinations of
matches (9 total) to ensure that we do not miss an data inserted in
incorrect order.

This commit also updates the find-patient directive to meet the styleguide
standards.