FIX(find-patient): remove SOUNDEX algorithm; fixes #708
I've updated the patient find directive to use strict matching on
first_name, last_name, middle_name.  We do all possible combinations of
matches (9 total) to ensure that we do not miss an data inserted in
incorrect order.

This commit also updates the find-patient directive to meet the styleguide
standards.


Former-commit-id: fc5dce969ea9b570b9b9c128a0b78f2c04288d54 [formerly a6df91276b62028860e2628462241a95a646ba5a]
Former-commit-id: b093719eda2bd9203d1712d4828eb151fa132490