FIX(distributions): patient distributions report
The patient distributions report has been consolidated into a
multi-report interface, where the user is able to specify a depot and a
report type.  The module will then include the report automatically as a
template, dropping in the required information.

This commit removes a large number of report/distribution routes that
are rendered unnecessary by these changes.