FIX(distributions): patient distributions report
The patient distributions report has been consolidated into a
multi-report interface, where the user is able to specify a depot and a
report type.  The module will then include the report automatically as a
template, dropping in the required information.

This commit removes a large number of report/distribution routes that
are rendered unnecessary by these changes.


Former-commit-id: fa7693eeb5b38fe98359a73161c51d20a2a71f1b [formerly bf714cc70737ecd1e031b5c9371cab84f9454bed]
Former-commit-id: f0fb6ea7eae1f2040f37d5ac932e70023ab3a217