FIX(expiring report): now accepts qs date params
This commit allows you to deep link the expiring report, as proposed by
issue #797.  Unfortunately, ng-model complains when you have a date of
00-00-0000, which is currently used to represent all expirations up to
present.  Therefore, some hacky code had to go into this to get it to
work properly.  If we are going to pursue this pattern, we ought to
figure out a better way of doing this.