FIX(sessions): Allow persistent sessions
This commit stops the server from deleting sessions from the file store.
That means a user must explicitly log out to remove their session
information (or the server must be restarted).

If we have better login error handling and messaging, this can be reset.
At the moment, it's very helpful so that users do not get kicked out of
their sessions while working.


Former-commit-id: 89c83b11a8918429c06081e212cc13715df63fcb [formerly f777c86dffad62fed8b612ffb583e5640f9ff71e]
Former-commit-id: 57884c20cb5c00e1f8de12bca2ccbf71d016a572