PATCH(depot loss): Posting support added
This commit fixes the posting journal route and allows the depot loss
controller to post to the journal using the request POST /depots/:uuid/distributions.

The next thing on the agenda is to calculate the correct value lost to
the enterprise.  Currently we simply pull the unit price out of the
inventory, when we really should look up the purchased price from the
purchase table.