CHORE(stock status): display stock projections
This commit removes the dependence on the fixed inventory numbers and
shows only the projected numbers from stock calculations over time.

NOTE that is change is hard-coded, and needs to provide more flexibility
to the users viewing the stock status report than it is able to provide
here.

Finally, the alert column is hidden in the printed form.