FIX(stock dashboard): proper loading indicators
This commit adds proper behavior to loading indicators on the stock
dashboard page, causing them to disappear if there is no data loaded.
We also add a message to the user, indicating that no data was found.

Finally, we've updated the SQL link to the stock management page, which
was recently broken.