FIX(stock dashboard): proper loading indicators
This commit adds proper behavior to loading indicators on the stock
dashboard page, causing them to disappear if there is no data loaded.
We also add a message to the user, indicating that no data was found.

Finally, we've updated the SQL link to the stock management page, which
was recently broken.


Former-commit-id: d9da309a52789749d7ec7cc38ca0440700d9d99b [formerly 35dcfe4b6886dbc2ec7f5e9dae7cacaecc4435d6]
Former-commit-id: 6185db2079a88aa4023d2029af827dcee229109c