FIX(trial balance): separate balance states
This commit changes the way error and results states work in the trial
balance.  Rather than having a single button with the message "view errors",
and is toggled _active_ when clicked, we now change the button text
depending on the current state.  While looking at results, it will still
read "view errors", but will switch to "view results" on click to
clearly indicate to the user to return.