FIX(trial balance): separate balance states
This commit changes the way error and results states work in the trial
balance.  Rather than having a single button with the message "view errors",
and is toggled _active_ when clicked, we now change the button text
depending on the current state.  While looking at results, it will still
read "view errors", but will switch to "view results" on click to
clearly indicate to the user to return.


Former-commit-id: c8afebf8a62239902e77d0537ae241aaf06d98f2 [formerly 6a94b7af5aaac816d1767ecb16c23528de6c3e2a]
Former-commit-id: cf39222b0e845ba83162ad9893b0dea547f2b06c