CHORE(partials): group partials by function
This commit imposes large structural changes on the bhima repository by
removing unused or old partials and grouping the rest logically by
function.  For example, the following top-level groups have been
created:
 1. payroll/
 2. employees/
 3. purchases/
To ensure that modules grouped in the tree are also grouped on the file
system of the server.

There may have been some minor route (URL) changes, which will be
addressed in a later commit.