FEAT(general ledger): optionally limit by date
This commit allows a user to limit the data loaded by the general_ledger
by date.  The default is to limit to the present date, only loaded
transactions made today.  Supported options are:
 1) Today
 2) This Week
 3) This Month
 4) All Time (all data)

This is a quick fix for Tshikaji.  Since our date component is now
fairly heavily used, it should be refactored into a service.