FEAT(exchange): rewrote exchange service
This commit adds a new exchange rate service that does not require the
use of a daily exchange rate.  Instead, we require that at least one
exchange rate is previously for each currency supported by the
enterprise.

The new exchange rate service is called ExchangeRateService and is found
in the file of the same name.  It uses intelligent caching to avoid
making unnecessary HTTP requests.  We expose the following API (subject
to change):
  - service.convertFromEnterpriseCurrency(currencyId, date, amount)
  - service.convertToEnbterpriseCurrency(currencyId, date, amount)

The new service also uses a server-side API to query the database.